+++ src/main/java/com/react/project2/dto/StudyDTO.java
@@ -79,8 +79,8 @@ public class StudyDTO {
                 .memberPhone(study.getMember().getPhone()) // 조회된 Member 엔티티를 사용합니다.
                 .location(study.getLocation())
                 .studyDeadlineDate(study.getStudyDeadlineDate())
-                .locationX((Double) study.getLocationX())
-                .locationY((Double) study.getLocationY())
+                .locationX(study.getLocationX())
+                .locationY(study.getLocationY())
                 .studyDate(study.getStudyDate())
                 .maxPeople(study.getMaxPeople())
                 .category(study.getCategory())
